Combine Interface: I/O; libpcu: repeated I/O
